+++ cmd/job/update.go
@@ -18,46 +18,46 @@ var newNameOpt = "new-name"
 
 var updateCmd = &cobra.Command{
 	Use:   "update",
-	Short: "Update cron",
+	Short: "Update job",
 	Run: func(cmd *cobra.Command, args []string) {
 		logger := slog.Default()
-		cronName := opts.GetStringOptOrExit(cmd, "name")
+		jobName := opts.GetStringOptOrExit(cmd, "name")
 		queries := config.GetDatabase(cmd.Context())
 
-		cron, err := queries.GetCron(cmd.Context(), cronName)
+		job, err := queries.GetJob(cmd.Context(), jobName)
 		if err != nil {
 			if err == sql.ErrNoRows {
-				core.LogErrorAndExit(logger, errors.New("cron with name '"+cronName+"' not found"))
+				core.LogErrorAndExit(logger, errors.New("job with name '"+jobName+"' not found"))
 			} else {
 				core.LogErrorAndExit(logger, err)
 			}
 		}
 
 		if cmd.Flags().Changed(notifyLogOpt) {
-			cron.Cron.NotifyLogContent = opts.GetBoolOptOrExit(cmd, notifyLogOpt)
+			job.Job.NotifyLogContent = opts.GetBoolOptOrExit(cmd, notifyLogOpt)
 		}
 		if cmd.Flags().Changed(newNameOpt) {
-			cron.Cron.Name = opts.GetStringOptOrExit(cmd, newNameOpt)
+			job.Job.Name = opts.GetStringOptOrExit(cmd, newNameOpt)
 		}
 
-		err = queries.UpdateCron(cmd.Context(), data.UpdateCronParams{
-			ID:               cron.Cron.ID,
-			Name:             cron.Cron.Name,
-			NotifyLogContent: cron.Cron.NotifyLogContent,
+		err = queries.UpdateJob(cmd.Context(), data.UpdateJobParams{
+			ID:               job.Job.ID,
+			Name:             job.Job.Name,
+			NotifyLogContent: job.Job.NotifyLogContent,
 		})
 
 		if err != nil {
-			core.LogErrorAndExit(logger, err, errors.New("unable to update cron"))
+			core.LogErrorAndExit(logger, err, errors.New("unable to update job"))
 		}
 
-		logger.Info("Cron updated")
+		logger.Info("Job updated")
 	},
 }
 
 func init() {
-	CronCmd.AddCommand(updateCmd)
-	updateCmd.Flags().String("name", "", "Cron Name (required)")
-	updateCmd.Flags().String(newNameOpt, "", "New cron Name")
+	JobCmd.AddCommand(updateCmd)
+	updateCmd.Flags().String("name", "", "Job Name (required)")
+	updateCmd.Flags().String(newNameOpt, "", "New job Name")
 	updateCmd.Flags().Bool(notifyLogOpt, false, "Includes the raw log output rather than the log filename in notification messages (default false)")
 	if err := updateCmd.MarkFlagRequired("name"); err != nil {
 		log.Fatalf("Unable to mark name as required %s", err)