+++ cmd/run/list.go
@@ -20,27 +20,27 @@ var listCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		logger := slog.Default()
 		conf := config.GetConfig()
-		cronName := opts.GetStringOptOrExit(cmd, nameOpt)
+		jobName := opts.GetStringOptOrExit(cmd, nameOpt)
 		queries := config.GetDatabase(cmd.Context())
 
-		if cronName != "" {
-			_, err := queries.GetCron(cmd.Context(), cronName)
+		if jobName != "" {
+			_, err := queries.GetJob(cmd.Context(), jobName)
 			if err != nil {
 				if err == sql.ErrNoRows {
-					core.LogErrorAndExit(logger, errors.New("cron with name '"+cronName+"' not found"))
+					core.LogErrorAndExit(logger, errors.New("job with name '"+jobName+"' not found"))
 				} else {
 					core.LogErrorAndExit(logger, err)
 				}
 			}
 		}
 
-		runRows, err := queries.GetRuns(cmd.Context(), cronName)
+		runRows, err := queries.GetRuns(cmd.Context(), jobName)
 		if err != nil {
 			core.LogErrorAndExit(slog.Default(), err)
 		}
 		tbl := table.New(
 			"ID",
-			"Cron Name",
+			"Job Name",
 			"Start Time",
 			"End Time",
 			"Log File",
@@ -50,7 +50,7 @@ var listCmd = &cobra.Command{
 		for _, run := range runRows {
 			tbl.AddRow(
 				run.Run.ID,
-				run.Cron.Name,
+				run.Job.Name,
 				core.FormatTime(run.Run.StartTime, conf.LocalTime),
 				core.FormatTime(run.Run.EndTime.Time, conf.LocalTime),
 				run.Run.LogFile,
@@ -66,5 +66,5 @@ var listCmd = &cobra.Command{
 func init() {
 	RunCmd.AddCommand(listCmd)
 
-	listCmd.Flags().String(nameOpt, "", "Name of cron to filter on")
+	listCmd.Flags().String(nameOpt, "", "Name of job to filter on")
 }