+++ main.go
@@ -8,8 +8,8 @@ import (
 	_ "embed"
 
 	"github.com/samcarswell/trochilus/cmd"
-	_ "github.com/samcarswell/trochilus/cmd/cron"
 	_ "github.com/samcarswell/trochilus/cmd/exec"
+	_ "github.com/samcarswell/trochilus/cmd/job"
 	_ "github.com/samcarswell/trochilus/cmd/run"
 )